[Feature][NativeModule][iOS][Part8] Initialize MTS NativeModule on iOS Platform
To align with Android, the enable_mts_module schema parameter controls the enabling of the entire MTS NativeModule.
On the iOS platform, a separate MainThreadModuleFactory is initialized to create the LynxModuleDarwin used in MTS.
The MainThreadModuleFactory initializes the LynxNativeModuleManager and further initializes the LynxLepusModuleManager, completing the creation of the entire MTS NativeModule.

issue: m-6654189539

doc: https://github.com/lynx-family/lynx/issues/1532

AutoSubmit: true